Add Table tests
