+++ tests/__init__.py
@@ -1,0 +1,8 @@
+table1_vals = [
+    ["a", "b", "c", "m", "r"],
+    ["d", None, "f", "m", "r"],
+    ["g", "h", None, "q", "r"],
+]
+table1_cols = [
+    "col0", "col1", "col2", "col3", "col4"
+]
